requirements.txt
torch
torchvision
scikit-learn
pandas
matplotlib
opencv-python
tqdm
PyYAML
